fix: 채용공고 여부 예외처리
채용공고가 없다면 ERROR

+++ services/recruit_service.js
@@ -1,5 +1,19 @@
 const {Recruit} = require('../models');
 
+const findRecruit = async (id) =>{
+    const recruit = await Recruit.findOne({
+        where: {
+            recruit_id:id
+        }
+    });
+    if(!recruit){
+        return true;
+    }
+    else {
+        return false;
+    }
+}
+
 const postNewRecruit = async (req,res,next)=>{
     // 채용공고 등록
     console.log('채용공고 등록');
@@ -15,6 +29,9 @@ const postNewRecruit = async (req,res,next)=>{
 
 const postUpdateRecruit = async (req,res,next)=>{
     // 채용공고 수정
+    if(findRecruit(req.body.recruit_id)){
+        return res.status(500).json({message : `채용공고 정보를 확인해 주세요`});
+    }
     await Recruit.update({
         position:req.body.position,
         reward: req.body.reward,
@@ -36,6 +53,9 @@ const postUpdateRecruit = async (req,res,next)=>{
 
 const postDeleteRecruit = async (req,res,next) => {
     // 채용공고 삭제
+    if(findRecruit(req.body.recruit_id)){
+        return res.status(500).json({message : `채용공고 정보를 확인해 주세요`});
+    }
     await Recruit.destroy({
         where: {
             recruit_id:req.body.recruit_id
